Separated connection service to base/standard/consumption format, fixed a connection reference issue for consumption
