+++ libs/services/designer-client-services/src/lib/base/index.ts
@@ -8,3 +8,7 @@ export { frequencyValues } from './manifests/schedule';
 // Connector
 export { BaseConnectorService } from './connector';
 export type { BaseConnectorServiceOptions } from './connector';
+// Connection
+export { BaseConnectionService } from './connection';
+export type { BaseConnectionServiceOptions } from './connection';
+export { IApiHubServiceDetails } from './connection';
