+++ libs/services/designer-client-services/src/lib/consumption/connection.ts
@@ -1,0 +1,3 @@
+import { BaseConnectionService } from '../base';
+
+export class ConsumptionConnectionService extends BaseConnectionService {}
