+++ libs/services/designer-client-services/src/lib/consumption/index.ts
@@ -1,3 +1,4 @@
 export { ConsumptionSearchService } from './search';
 export { ConsumptionOperationManifestService } from './operationmanifest';
 export { ConsumptionConnectorService } from './connector';
+export { ConsumptionConnectionService } from './connection';
