+++ libs/services/designer-client-services/src/lib/standard/index.ts
@@ -1,4 +1,4 @@
-export { StandardConnectionService, IApiHubServiceDetails } from './connection';
+export { StandardConnectionService } from './connection';
 export { StandardConnectorService } from './connector';
 export { StandardGatewayService } from './gateway';
 export { StandardOperationManifestService } from './operationmanifest';
